fix: issue #1
In the default environment the required path ```/usr/local/python/current/bin:/usr/local/py-utils/bin:/usr/local/share/nvm/current/bin:/usr/local/bin``` is out of order in the $PATH execution order.  While it is included in the $PATH variable, it does not proceed the conda/miniconda paths and seems to create underlying dependency issue with Python required C++ and possibly other files on the local system.

This is only affecting non-login shells.  If you manually (or via scripted args) force an interactive login shell (e.g. ```bash -l```) then the correct/preferred path order is set.

Explicitly setting the required path order in ~/.bashrc should fix this. for interactive and non-interactive shells a like. 

Note: while this should fix the issue, it should not be required or a deeper analysis of how $PATH should be loaded needs to be examined in the future.